code restructure and added dataset based rms reject function
